Smoketest - find a free port for geth
Find & check a free TCP port that will be later used by
geth to bind the RPC interface.
